Now Finished for 3D L System and  Texture Mapping
But there is a bug when you first open the project, which is there will be nothing show up. You should uncomment the fragment shader first and then restore it, after which it will become normal.

But I am still trying to find a better solution about L System rules. Our one is too ugly.