Update and Some new functions
New functions added to the project! Include thickness change when the tree grows, making the tree more real. Meanwhile, particle system is added to the project to simulate snow. (Please do try to set near plane to 1.0 and open the snow effect, you can see really beautiful stuff. Fianlly, 'Forest' now is able to be created, which will be much more better if Depth of Field is implemented!